removed the pb model loader
