+++ src/executor/frozen_model_test.ts
@@ -23,13 +23,9 @@ import * as fm from './frozen_model';
 
 const HOST = 'http://example.org';
 const TFHUB_SUFFIX = '?tfjs-format=file';
-const MODEL_URL = `${HOST}/model.pb`;
-const WEIGHT_MANIFEST_URL = `${HOST}/weights_manifest.json`;
+const MODEL_URL = `${HOST}/model.json`;
 const RELATIVE_MODEL_URL = '/path/model.pb';
-const RELATIVE_WEIGHT_MANIFEST_URL = '/path/weights_manifest.json';
-const TFHUB_MODEL_URL = `${HOST}/model/1/tensorflowjs_model.pb${TFHUB_SUFFIX}`;
-const TFHUB_WEIGHT_MANIFEST_URL =
-    `${HOST}/model/1/weights_manifest.json${TFHUB_SUFFIX}`;
+const TFHUB_MODEL_URL = `${HOST}/model/1/model.json${TFHUB_SUFFIX}`;
 let model: fm.FrozenModel;
 const bias = tfc.tensor1d([1], 'int32');
 const OCTET_STREAM_TYPE = 'application/octet-stream';
@@ -108,35 +104,26 @@ const DYNAMIC_SHAPE_MODEL: tensorflow.IGraphDef = {
 };
 describe('Model', () => {
   beforeEach(() => {
-    model = new fm.FrozenModel(MODEL_URL, WEIGHT_MANIFEST_URL);
-    spyOn(window, 'fetch').and.callFake(async (path: string) => {
-      if (path === MODEL_URL || path === RELATIVE_MODEL_URL ||
-          path === TFHUB_MODEL_URL) {
-        return new Response(
-            new Uint8Array([1, 2, 3]),
-            {'headers': {'Content-Type': OCTET_STREAM_TYPE}});
-      } else if (
-          path === WEIGHT_MANIFEST_URL ||
-          path === RELATIVE_WEIGHT_MANIFEST_URL ||
-          path === TFHUB_WEIGHT_MANIFEST_URL) {
-        return new Response(
-            JSON.stringify(weightsManifest),
-            {'headers': {'Content-Type': JSON_TYPE}});
-      } else if (
-          path.match(`${HOST}.*/weight_0`) || path === '/path/weight_0') {
-        return new Response(
-            bias.dataSync() as Int32Array,
-            {'headers': {'Content-Type': OCTET_STREAM_TYPE}});
-      } else {
-        throw new Error(`Invalid path: ${path}`);
-      }
-    });
+    model = new fm.FrozenModel(MODEL_URL);
   });
-  afterEach(() => {});
 
   describe('simple model', () => {
     beforeEach(() => {
-      spyOn(tensorflow.GraphDef, 'decode').and.returnValue(SIMPLE_MODEL);
+      spyOn(window, 'fetch').and.callFake(async (path: string) => {
+        if (path === MODEL_URL || path === RELATIVE_MODEL_URL ||
+            path === TFHUB_MODEL_URL) {
+          return new Response(
+              JSON.stringify({modelTopology: SIMPLE_MODEL, weightsManifest}),
+              {'headers': {'Content-Type': JSON_TYPE}});
+        } else if (
+            path.match(`${HOST}.*/weight_0`) || path === '/path/weight_0') {
+          return new Response(
+              bias.dataSync() as Int32Array,
+              {'headers': {'Content-Type': OCTET_STREAM_TYPE}});
+        } else {
+          throw new Error(`Invalid path: ${path}`);
+        }
+      });
     });
 
     it('load', async () => {
@@ -232,7 +219,7 @@ describe('Model', () => {
     describe('dispose', () => {
       it('should dispose the weights', async () => {
         const numOfTensors = tfc.memory().numTensors;
-        model = new fm.FrozenModel(MODEL_URL, WEIGHT_MANIFEST_URL);
+        model = new fm.FrozenModel(MODEL_URL);
 
         await model.load();
         model.dispose();
@@ -250,8 +237,7 @@ describe('Model', () => {
 
     describe('relative path', () => {
       beforeEach(() => {
-        model = new fm.FrozenModel(
-            RELATIVE_MODEL_URL, RELATIVE_WEIGHT_MANIFEST_URL);
+        model = new fm.FrozenModel(RELATIVE_MODEL_URL);
       });
 
       it('load', async () => {
@@ -261,18 +247,14 @@ describe('Model', () => {
     });
 
     it('should loadFrozenModel', async () => {
-      const model = await fm.loadFrozenModel(MODEL_URL, WEIGHT_MANIFEST_URL);
+      const model = await fm.loadFrozenModel(MODEL_URL);
       expect(model).not.toBeUndefined();
     });
 
     it('should loadFrozenModel with request options', async () => {
-      const model = await fm.loadFrozenModel(
-          MODEL_URL, WEIGHT_MANIFEST_URL, {credentials: 'include'});
+      const model =
+          await fm.loadFrozenModel(MODEL_URL, {credentials: 'include'});
       expect(window.fetch).toHaveBeenCalledWith(MODEL_URL, {
-        credentials: 'include',
-        headers: Object({Accept: OCTET_STREAM_TYPE})
-      });
-      expect(window.fetch).toHaveBeenCalledWith(WEIGHT_MANIFEST_URL, {
         credentials: 'include',
         headers: Object({Accept: JSON_TYPE})
       });
@@ -303,7 +285,22 @@ describe('Model', () => {
 
   describe('control flow model', () => {
     beforeEach(() => {
-      spyOn(tensorflow.GraphDef, 'decode').and.returnValue(CONTROL_FLOW_MODEL);
+      spyOn(window, 'fetch').and.callFake(async (path: string) => {
+        if (path === MODEL_URL || path === RELATIVE_MODEL_URL ||
+            path === TFHUB_MODEL_URL) {
+          return new Response(
+              JSON.stringify(
+                  {modelTopology: CONTROL_FLOW_MODEL, weightsManifest}),
+              {'headers': {'Content-Type': JSON_TYPE}});
+        } else if (
+            path.match(`${HOST}.*/weight_0`) || path === '/path/weight_0') {
+          return new Response(
+              bias.dataSync() as Int32Array,
+              {'headers': {'Content-Type': OCTET_STREAM_TYPE}});
+        } else {
+          throw new Error(`Invalid path: ${path}`);
+        }
+      });
     });
 
     it('should throw error if call predict directly', async () => {
@@ -337,7 +334,22 @@ describe('Model', () => {
 
   describe('dynamic shape model', () => {
     beforeEach(() => {
-      spyOn(tensorflow.GraphDef, 'decode').and.returnValue(DYNAMIC_SHAPE_MODEL);
+      spyOn(window, 'fetch').and.callFake(async (path: string) => {
+        if (path === MODEL_URL || path === RELATIVE_MODEL_URL ||
+            path === TFHUB_MODEL_URL) {
+          return new Response(
+              JSON.stringify(
+                  {modelTopology: DYNAMIC_SHAPE_MODEL, weightsManifest}),
+              {'headers': {'Content-Type': JSON_TYPE}});
+        } else if (
+            path.match(`${HOST}.*/weight_0`) || path === '/path/weight_0') {
+          return new Response(
+              bias.dataSync() as Int32Array,
+              {'headers': {'Content-Type': OCTET_STREAM_TYPE}});
+        } else {
+          throw new Error(`Invalid path: ${path}`);
+        }
+      });
     });
 
     it('should throw error if call predict directly', async () => {
