+++ tslint.json
@@ -2,7 +2,7 @@
   "extends": ["tslint-no-circular-imports"],
   "linterOptions": {
     "exclude": [
-        "src/data/compiled_api_json.ts"
+        "src/data/compiled_api.ts"
     ]
   },
   "rules": {
